+++ src/main/java/com/learn/lld/designPrincipal/level1/payment/PaymentInterface.java
@@ -1,7 +1,5 @@
 package com.learn.lld.designPrincipal.level1.payment;
 
-import java.math.BigDecimal;
-
 public interface PaymentInterface {
-    public String processPayment(BigDecimal amount);
+    public String processPayment(double amount);
 }
